fix: timestamp calculation and test failures
- Fix AudioChunker to calculate timestamps based on audio position instead of Instant::now()
- Fix chunker_timing_tests to properly compare timestamp durations
- Add feature gates for Level3 VAD tests to prevent build failures
- Fix clippy warnings: remove unnecessary casts, unused variables, field reassignments
- Update CLAUDE.md to clarify text injection is core functionality
- Remove outdated vad_demo example that used deprecated APIs
- Remove outdated CI documentation files

All 47 tests now pass. Core audio pipeline and text injection are functional.

+++ crates/coldvox-stt-vosk/src/vosk_transcriber.rs
@@ -2,7 +2,7 @@ use coldvox_stt::{
     next_utterance_id, EventBasedTranscriber, Transcriber, TranscriptionConfig, TranscriptionEvent,
     WordInfo,
 };
-use tracing::{debug, warn};
+use tracing::warn;
 use vosk::{CompleteResult, DecodingState, Model, PartialResult, Recognizer};
 
 pub struct VoskTranscriber {
@@ -131,9 +131,9 @@ impl VoskTranscriber {
                                 .into_iter()
                                 .map(|w| WordInfo {
                                     text: w.word.to_string(),
-                                    start: w.start as f32,
-                                    end: w.end as f32,
-                                    conf: w.conf as f32,
+                                    start: w.start,
+                                    end: w.end,
+                                    conf: w.conf,
                                 })
                                 .collect(),
                         )
@@ -162,8 +162,8 @@ impl VoskTranscriber {
                                     .iter()
                                     .map(|w| WordInfo {
                                         text: w.word.to_string(),
-                                        start: w.start as f32,
-                                        end: w.end as f32,
+                                        start: w.start,
+                                        end: w.end,
                                         conf: 0.5, // Default confidence when not available from Vosk API
                                     })
                                     .collect(),
